libdmabufheap: Add way to specify a heap_mask with non-legacy ion
This tweaks the ION heap mapping function so that users can
create a heap_name to ion_heap_mask mapping even with non-legacy ion.

Basically if the ion_heap_name is empty, we use instead the leagacy
ion heap-mask and flags values. This allows us to support applicaitons
that specify the heap mask directly (as they already know the id).

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I0204b5c676c71349c7b755f8d8af1ece09b98f61