libdmabufheap: Fix build issues with c-style cpp code using the wrapper header
When initially porting the hikey960 gralloc to use libdmabufheap,
the code was written in C, however the filename was a .cpp file.

This tripped over build issues trying to use the wrapper header,
where the BufferAllocator typedef wasn't defined.

This patch resolves the issue by adding a typedef for class BufferAllocator
as well, which resolves the issue.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I4df0b83d638bc98c0f4119eaa250384ca0b65d1b