libdmabufheap: Provide a way to better check that we have a valid name mapping
The heap name mapping from dmabuf heaps and ION only returns
error if the ion heap name doesn't exist or if its a legacy ion
system and the legacy heap mask is null.

This makes it difficult to detect when calling the mapper function
to see if we have a valid name mapping for a given heap name.

For instance, if there is a name mapping for:
  (foo) => (ION_FOO, 0, ION_FOO_MASK, 0)

On modern ION systems, "foo" will map to ION_FOO heap
On dmabuf heap systems, "foo" will map to /dev/dma_heap/foo

But there may be systems where we are conditionally creating
a name mapping for a heap name that may not exist, and it
would be helpful to be able to detect that.

The current code doesn't allow it, so this patch changes
the name mapping function to try to open the heap name
as a DMABUF heap initially. If it finds such a heap, it
doesn't have to bother with the ION name mappings.

Only if there is no DMABUF Heap with the specified name
will it create the ION name mappings, which can also fail
if the heap is not supported.

This allows us to get a proper error when creating a name
mapping, when there is no heap which supports the mapping.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I3341e5d5200c04eecda5a4236e467910f96cf037