libdmabufheap: Add an API to check ION support
Although libdmabufheap is intended to be an abstraction layer that is
intended to work with both ION and DMA-BUF heaps, there are times when
its clients would like to differentiate between the devices supporting
ION and DMA-BUF heaps.

This API returns if a device supports ION by checking for the presence
of /dev/ion.

Test: atest dmabuf_heap_test
Bug: 183481820
Change-Id: I0f039826431c6e11b76d35253eca6b6ed31e0a89