libdmabufheap: Set heap name as dmabuf name
libdmabufheap does not set a name for dmabufs that it allocates. This
causes most buffers to show up with "<unknown>" names in dmabuf_dump.
This change sets the name to the name of the heap the buffer came from.
Note that DMA_BUF_NAME_LEN is currently 32 but dmabuf heaps do not have
a name length limit, so this operation will fail if the heap name is too
long. This also adds the heap name to dmabuf entries in the
/proc/pid/maps file. End-users of the buffer are still free to overwrite
this name with something more specific to the use-case.

Test: raven:/ # dmabuf_dump
Test: oid.apps.photos:17224
Test:                  Name              Rss              Pss         nr_procs            Inode
Test:       system-uncached           392 kB           392 kB                1             2185
Test:                system             4 kB             4 kB                1             2186
Test:         PROCESS TOTAL           396 kB           396 kB
Test: raven:/ # cat /proc/2083/maps|grep dmabuf
Test: 6fd53eb000-6fd53ec000 rw-s 00000000 00:09 168                            /dmabuf:system
Test: 6fd53ec000-6fd53ed000 rw-s 00000000 00:09 166                            /dmabuf:system

Change-Id: Ied8ba526d5e88fb6860285b6e7c6027f9a39e80a