libdmabufheap: Remove most ION support
All kernels with ION support will reach end of life in December 2025,
and ION will not be supported in any future Android release.

Because this is a vendor available library, we must maintain a stable
ABI for GRF. Unfortunately this library was exposed with a C++ ABI. This
means the private ION specific class members must be retained even
though they are no longer used.

go/android-ndk-api-guidelines#headers

Bug: 444679893
Change-Id: Iadd01b809dc94737ebeef077e89f3ed14e015e3f